Shell lele lilili

+++ src/lib-header/framebuffer.h
@@ -29,7 +29,7 @@
  */
 void framebuffer_write(uint8_t row, uint8_t col, char c, uint8_t fg, uint8_t bg);
 
-void framebuffer_write_buf(char * buf, uint8_t len, uint8_t color);
+void framebuffer_write_buf(char * buf, uint32_t len, uint8_t color);
 
 /**
  * Set cursor to specified location. Row and column starts from 0
